Adds styling and HTML markup
Experiments with markup and layout, as well as colour and font styling.

Comments are yet to be added, and it is understood that the structure
of the CSS file (order of rules) could be improved.

Knowing this is only a test, colours have been intentionally made a
little odd as a creative test (these would not pass WCAG colour
compliance).